config: move set sda scl to board init
In the process of merging to linorobot/linorobot2_hardware upstream,
the changes to main firmware.ino should be minimized. So the setting
of specific sda scl is moved to board config.
If you use specific i2c pins, please update BOARD_INIT to do this in
your board config accordingly. See esp32_config.h in my master branch
as an example.